Ajout de la classe WaveSpawner pour séparer le spawn du GameManager
