+++ Assets/Prefabs/GameManager.prefab
@@ -10,6 +10,7 @@ GameObject:
   m_Component:
   - component: {fileID: 6068187390962822522}
   - component: {fileID: 9222688913029514041}
+  - component: {fileID: 6068187392579199654}
   m_Layer: 0
   m_Name: GameManager
   m_TagString: Untagged
@@ -43,10 +44,20 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: a0e92472b221f88418cb12f70fba0eaf, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  restartScene: 
+  restartScene: EnemyScene
   endGameOverlay: {fileID: 0}
-  enemies:
-  - {fileID: 6775578803265522924, guid: 7776fd347df61474e8bd30f647421ebc, type: 3}
-  - {fileID: 2246748386601214517, guid: 970ba45ba63d04f49b08e3824a6575d3, type: 3}
-  - {fileID: 7507134723547207384, guid: 6d16da6939364f74cbf3e0441aaa250a, type: 3}
-  player: {fileID: 565854060559499053, guid: 3bcedb26cd32dd64ab294bbafaad6d50, type: 3}
+--- !u!114 &6068187392579199654
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 6068187390962822520}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 872ec3f559c24b64686d8415e63af725, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  player: {fileID: 0}
+  enemiesType: []
+  waves: []
