+++ Assets/Scripts/Enemy/Wave.cs
@@ -1,0 +1,11 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class Wave : MonoBehaviour
+{
+    public int nbEnemies;
+    public float waveTime;
+    
+
+}
