+++ Assets/Scripts/Enemy/Wave.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d9356ac6b1c7bb74d9834f6edfc97127
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
