+++ Assets/Scripts/Enemy/WaveSpawner.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 47208c1c01655664baf0b81f2a6e6075
+guid: 872ec3f559c24b64686d8415e63af725
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
