+++ Assets/Scripts/GameManager.cs
@@ -6,13 +6,9 @@ using UnityEngine.SceneManagement;
 public class GameManager : MonoBehaviour
 {
     public static GameManager instance = null;
-    private EnemiesFactory enemyFactory;
-    private float waveTimer = 0f;
+
     [SerializeField] private string restartScene;
     [SerializeField] private GameObject endGameOverlay;
-    [SerializeField] private List<Enemy> enemies;
-
-    [SerializeField] private GameObject player;
 
     Dictionary<HealthBar.Color, int> spawns = new Dictionary<HealthBar.Color, int>();
     Dictionary<HealthBar.Color, float> ratios = new Dictionary<HealthBar.Color, float>();
@@ -24,8 +20,6 @@ public class GameManager : MonoBehaviour
         if (instance == null)
         instance = this;
 
-        player = GameObject.Find("FeuFollet");
-        enemyFactory = new EnemiesFactory(enemies);
         
         spawns.Add(HealthBar.Color.RED, 0);
         spawns.Add(HealthBar.Color.BLUE, 0);
@@ -45,7 +39,6 @@ public class GameManager : MonoBehaviour
         
         if(endGameOverlay == null)
             endGameOverlay = GameObject.Find("Overlay");
-        enemyFactory.GenerateEnemies(50);
     }
 
     void StartMusic()
@@ -58,53 +51,21 @@ public class GameManager : MonoBehaviour
         am.setSoundVolume("StemRed", 0f);
         am.setSoundVolume("StemGreen", 0f);
         am.setSoundVolume("StemBlue", 0f);
-    }
+    }    
 
-    void Update() 
+    public void SpawnEnemy(HealthBar.Color enemyType)
     {
-        if(player.activeSelf)
-        {
-            if(!enemyFactory.IsEmpty())
-            SpawnEnemy();
-            else if(waveTimer >= 7)
-            {
-                waveTimer = 0;
-                enemyFactory.GenerateEnemies(40);        
-            }
-        }
-
-        //UpdateMainTheme();
-    }
-
-    void FixedUpdate()
-    {
-        waveTimer += Time.deltaTime; 
-    }
-
-    void SpawnEnemy()
-    {
-        Enemy nextEnemy = enemyFactory.NextEnemy();
-        //Debug.LogWarning("NEXT ENEMY TYPE : " + nextEnemy.GetEnemyDamageType());
+        spawns[enemyType] += 1;
         
-        spawns[nextEnemy.GetEnemyDamageType()] += 1;
-        //spawns[nextEnemy.GetEnemyDamageType()] = spawns[nextEnemy.GetEnemyDamageType()] + 1;
-        
-        Instantiate(nextEnemy, RandomPos(), Quaternion.identity);
         CalculateRatios();
         UpdateMainTheme();
-        
     } 
 
     private void UpdateMainTheme()
     {
-        //Debug.Log("UPDATING WITH RED : " + spawns[HealthBar.Color.RED]);
-        //Debug.Log("UPDATING WITH BLUE : " + spawns[HealthBar.Color.BLUE]);
         am.setSoundVolume("StemRed", ratios[HealthBar.Color.RED]);
         am.setSoundVolume("StemGreen", ratios[HealthBar.Color.GREEN]);
         am.setSoundVolume("StemBlue", ratios[HealthBar.Color.BLUE]);
-        Debug.Log("RED VOLUME : " + am.getSoundVolume("StemRed"));
-        Debug.Log("BLUE VOLUME : " + am.getSoundVolume("StemBlue"));
-        Debug.Log("GREEN VOLUME : " + am.getSoundVolume("StemGreen"));
     }
 
     public void DecrementEnemyCount(HealthBar.Color enemyType)
@@ -117,7 +78,6 @@ public class GameManager : MonoBehaviour
     private void CalculateRatios()
     { 
         float total = spawns[HealthBar.Color.BLUE] + spawns[HealthBar.Color.RED] + spawns[HealthBar.Color.GREEN];
-        //Debug.Log("TOTAL ENEMIES : " + total);
         if(total == 0f)
             return;
         ratios[HealthBar.Color.BLUE] = (float)spawns[HealthBar.Color.BLUE]/total;
@@ -125,15 +85,7 @@ public class GameManager : MonoBehaviour
         ratios[HealthBar.Color.GREEN] = (float)spawns[HealthBar.Color.GREEN]/total;      
     }
 
-    protected Vector3 RandomPos() 
-    {
-        float radius = Random.Range(15f, 30f);
-        float angle = Random.Range(0f, 360f);
-        Vector3 newPos = new Vector3();
-        newPos.y = player.transform.position.y + (radius * Mathf.Sin(angle));
-        newPos.x = player.transform.position.x + (radius * Mathf.Cos(angle));
-        return newPos;
-    }
+    
 
     public void StartGame()
     {
